network_services/grayscaler/requirements.txt
fastapi
uvicorn
requests
pillow
python-multipart